refactor: optimize loading and chunck components
refactor: optimize loading and chunck components
