--- features/home/index.ts
@@ -1,2 +1,0 @@
-export { AuthorHeader } from './author-header';
-export { AuthorContent } from './author-content';
