+++ pages/blog/article/[slug].tsx
@@ -1,22 +1,20 @@
 /* eslint-disable no-tabs */
 import { useRouter } from 'next/router';
 import ARTICLE_QUERY from 'queries/article/article';
-import markdownLight from 'styles/github-markdown-css-light.module.css';
-import { AiOutlineArrowLeft, AiOutlineCalendar } from 'react-icons/ai';
-import { RiTimer2Line } from 'react-icons/ri';
 import * as CSS from 'features/article/styles';
 import type { GetStaticPaths, GetStaticProps } from 'next';
 import { ARTICLES_QUERY } from 'queries/articles/articles';
 import apolloClient from 'utils/apollo-client';
-import { getReadingTime } from 'utils/getTimeReading';
-import { getLocaleDate } from 'utils/get-locale-date';
 import { NextSEO } from 'components/SEO';
-import Link from 'next/link';
 import { ArticleLayout } from 'features/ui/layouts';
 import { blurImage } from 'utils/blur-image';
 import { ArticleData } from 'features/article/ts';
 import { ArticleFooter } from 'features/article';
 import JSONLD from 'components/JSON-LD';
+import dynamic from 'next/dynamic';
+
+const ArticleHeader = dynamic(() => import('features/article/header'), { ssr: true });
+const Markdown = dynamic(() => import('features/article/markdown'), { ssr: true });
 
 const ArticleItem = ({ articles, blurDataURL }: ArticleData) => {
   if (!articles) return null;
@@ -24,8 +22,6 @@ const ArticleItem = ({ articles, blurDataURL }: ArticleData) => {
   const router = useRouter();
 
   const [article] = articles;
-  const { readTime } = getReadingTime(article?.attributes?.content);
-  const { localeDate: publishedAt } = getLocaleDate(article?.attributes?.publishedAt, 'pt-BR');
 
   return (
     <>
@@ -48,35 +44,7 @@ const ArticleItem = ({ articles, blurDataURL }: ArticleData) => {
       <ArticleLayout>
         <CSS.ArticleContainer>
           <CSS.ArticleMarkdownContainer>
-            <CSS.ArticleHeader>
-              <div>
-                <CSS.InfoHeader>
-                  <Link href="/blog">
-                    <CSS.BackToOverview
-                      type="button"
-                      aria-label="Back to overview"
-                    >
-                      <AiOutlineArrowLeft size={19} aria-hidden="true" />
-                      <p>Back to overview</p>
-                    </CSS.BackToOverview>
-                  </Link>
-                  <div>
-                    <CSS.DateTimeRead>
-                      <AiOutlineCalendar size={17} />
-                      {publishedAt}
-                      &nbsp;|&nbsp;
-                      <RiTimer2Line size={17} />
-                      {readTime}
-                    </CSS.DateTimeRead>
-                  </div>
-                </CSS.InfoHeader>
-              </div>
-
-              <CSS.ArticleTitle>{article?.attributes?.title}</CSS.ArticleTitle>
-              <CSS.ArticleDescription>
-                {article?.attributes?.description}
-              </CSS.ArticleDescription>
-            </CSS.ArticleHeader>
+            <ArticleHeader article={article} />
             <CSS.ImageContainer>
               <CSS.ArticleImage
                 width={5000}
@@ -91,11 +59,7 @@ const ArticleItem = ({ articles, blurDataURL }: ArticleData) => {
                 {article.attributes.image.data.attributes.caption}
               </CSS.ImageCredit>
             </CSS.ImageContainer>
-            <CSS.Article>
-              <CSS.ArticleMarkdown className={markdownLight['markdown-body']}>
-                {article?.attributes?.content}
-              </CSS.ArticleMarkdown>
-            </CSS.Article>
+            <Markdown article={article} />
           </CSS.ArticleMarkdownContainer>
           <ArticleFooter
             author={article?.attributes?.author?.data?.attributes?.name}
